refactor(services): reorganize directory structure to "Flat-to-Modular"
- Sub-categorize src/services into data-ingestion, validation, and broadcast.
- Update 120+ import paths across the repository to reflect the new structure.
- Fix internal relative imports for relocated service files.
- Maintain core utility services in the root of src/services.

Closes #259
