+++ src/services/data-ingestion/sorobanEventListener.ts
@@ -1,11 +1,11 @@
 import { Horizon } from "@stellar/stellar-sdk";
 // import type { OnChainPrice } from "@prisma/client";
 import type { ServerApi } from "@stellar/stellar-sdk/lib/horizon";
-import prisma from "../lib/prisma";
-import { broadcastToSessions } from "../lib/socket";
-import stellarProvider from "../lib/stellarProvider";
+import prisma from "../../lib/prisma";
+import { broadcastToSessions } from "../../lib/socket";
+import stellarProvider from "../../lib/stellarProvider";
 import dotenv from "dotenv";
-import { signer } from "../signer";
+import { signer } from "../../signer";
 
 dotenv.config();
 
@@ -314,3 +314,4 @@ export class SorobanEventListener {
     return this.oraclePublicKey;
   }
 }
+
