+++ src/services/validation/priceProtection.ts
